app_ab/requirements.txt
# Dependências Python para gerenciar Supabase

# Supabase client
supabase==2.10.0

# PostgreSQL adapter
psycopg2-binary==2.9.10

# Environment variables
python-dotenv==1.0.1

# Data manipulation (opcional)
pandas==2.2.3

# HTTP requests (opcional)
requests==2.32.3
